Initialized WebUI project with MudBlazor

+++ src/Application/Entities/Customers/Queries/GetCustomerById/CustomerDto.cs
@@ -7,7 +7,7 @@ public class CustomerDto : IMapFrom<Customer>
     public string LastName { get; set; }
     public string Email { get; set; }
     public DateTime DateOfBirth { get; set; }
-    public byte PhoneCountryCode { get; set; }
+    public short PhoneCountryCode { get; set; }
     public long PhoneNumber { get; set; }
     public string BankAccountNumber { get; set; }
 }
