+++ src/Domain/Entities/Customer.cs
@@ -15,7 +15,7 @@ public class Customer : BaseEntity
     }
 
     public DateTime DateOfBirth { get; set; }
-    public byte PhoneCountryCode { get; set; }
+    public short PhoneCountryCode { get; set; }
     public long PhoneNumber { get; set; }
     public string BankAccountNumber { get; set; }
 }
